requirements.txt
asgiref==3.8.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dataclasses==0.6
Django==5.0.6
djangorestframework==3.15.1
drf-yasg==1.21.7
fonttools==4.44.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-speech==2.22.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
idna==3.4
inflection==0.5.1
kiwisolver==1.4.5
matplotlib==3.8.1
numpy==1.26.1
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydub==0.25.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scipy==1.11.3
six==1.16.0
soundfile==0.12.1
SpeechRecognition==3.10.0
sqlparse==0.5.0
srt==3.5.3
torchlibrosa==0.0.1
tqdm==4.66.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vosk==0.3.45
Wave==0.0.2
websockets==12.0
